Save history into csv. Improve last audio removal. Add a DEV mode with the app

+++ flask_app.py
@@ -7,6 +7,8 @@ from flask import Flask, request, render_template, jsonify, url_for, session
 import openai
 from gtts import gTTS
 import secrets
+import csv
+import uuid
 
 # Loading OpenAI API key from configuration file
 config = configparser.ConfigParser()
@@ -14,7 +16,9 @@ config.read('config.ini')
 openai.api_key = config.get('OPENAI_API', 'key')
 
 # Development mode, unable requests to OpenAI
-DEV_MODE = True
+DEV_MODE = False
+# Used when testing the app with the flask server locally, needs a signed certificate
+DEV_MODE_APP = False
 
 # Initializing Flask app
 app = Flask(__name__)
@@ -37,7 +41,7 @@ def transcribe_audio():
 
     # Securing the filename and saving it in the defined upload directory
     audio_file = request.files['audio']
-    filename = f"{secure_filename(audio_file.filename)}.wav"
+    filename = f"{secure_filename(audio_file.filename)}_{uuid.uuid4()}.wav"
     file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
     audio_file.save(file_path)
 
@@ -51,7 +55,15 @@ def transcribe_audio():
         transcript = "This is DEV mode."
     else:
         transcript = transcribe(file_path, input_language)
+
+    # Save the transcript to the CSV file along with the IP address and User-Agent
+    user_agent = request.headers.get('User-Agent', 'Unknown')  # Default to 'Unknown' if User-Agent header is missing
+    save_to_csv(transcript, request.remote_addr, user_agent)
     
+    # Remove recording
+    if os.path.exists(file_path):
+        os.remove(file_path)
+        
     return jsonify({'transcript': transcript})
 
 @app.route('/translate', methods=['POST'])
@@ -86,7 +98,8 @@ def translate_audio():
 
         # Remove the previous audio file
         if not session.get('last_audio_file', None) == None:
-            os.remove(session.get('last_audio_file', ''))
+            if os.path.exists(session.get('last_audio_file', '')):
+                os.remove(session.get('last_audio_file', ''))
             
         # Saving the speech file to the audio directory
         timestamp = datetime.now().strftime("%Y%m%d%H%M%S")
@@ -167,6 +180,22 @@ def wait_for_file(file_path):
     while not os.path.exists(file_path) or not os.path.getsize(file_path) > 0:
         time.sleep(0.1)
 
+def save_to_csv(transcript, ip_address, user_agent, filename="history/transcripts.csv"):
+    # Check if the directory exists, if not, create it
+    directory = os.path.dirname(filename)
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+
+    with open(filename, mode='a', newline='') as file:
+        writer = csv.writer(file)
+        # The current time, transcript, IP address, and User-Agent are saved
+        writer.writerow([datetime.now(), transcript, ip_address, user_agent])
+
+
+
 # Run the Flask app
 if __name__ == '__main__':
-    app.run(debug=True, host="0.0.0.0", port=5009)
+    if DEV_MODE_APP:
+        app.run(ssl_context=('cert.pem', 'key.pem'), debug=True, host="0.0.0.0", port=5009)
+    else:
+        app.run(host="0.0.0.0", port=5009)
